requirements.txt
alembic==1.5.5
aniso8601==9.0.0
Authlib==0.15.3
cachelib==0.1.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.6
ecdsa==0.16.1
Flask==1.0.3
Flask-Cors==3.0.7
Flask-Migrate==2.6.0
Flask-Moment==0.11.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.3
future==0.18.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.4
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.3.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-jose==3.2.0
python-jose-cryptodome==1.3.2
pytz==2021.1
requests==2.25.1
rsa==4.7.2
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.3.4
urllib3==1.26.3
Werkzeug==1.0.1
WTForms==2.3.3
